# JDK-8221691: CSS font caching could be improved