@kevinrushforth 2019-04-02
CSS is a very complex area, and any performance optimization runs the very real risk of introducing a regression (we have several examples where this has happened). This is especially true when making changes that avoid redundant work. The challenge is to prove that the skipped computation is actually unneeded in all cases. I presume you have run all of the existing test cases? A full, automated test run, including robot tests, plus a manual run of Ensemble8 testing various apps would be a minimum.

Before we review the code, please provide a self-contained (in this PR) evaluation of the performance problem and your solution. Also, I would like to see a discussion of the trade-offs and risks associated with doing this.

In particular:

1. What sort of regression testing is needed? Think about what could go wrong and then write a test for it. @dgrieve gave a couple pointers in [this message](https://mail.openjdk.java.net/pipermail/openjfx-dev/2019-January/023076.html) on the openjfx-dev alias.

2. What use cases will benefit from the performance improvement? As part of this, please provide at least one performance test program that shows a before / after comparison.