@kevinrushforth 2019-04-02
@dgrieve your input would be welcome if you have a few minutes to take a look at this and comment on it.