@dgrieve 2019-04-03
Will do.

On 4/2/19 4:10 PM, Kevin Rushforth wrote:
>
> @dgrieve 
> <https://urldefense.proofpoint.com/v2/url?u=https-3A__github.com_dgrieve&d=DwMCaQ&c=RoP1YumCXCgaWHvlZYR8PZh8Bv7qIrMUB65eapI_JnE&r=H6jKHjHoLf6vCNF0PYEnJ6cS6kScbGLraJqgoRs9znU&m=TMsfvQuwAKRQX0c5NNEKf5Oo1Fpp1ZYnCUAIfdHcnTs&s=t3_uBuMpPDR7aqbj7sFR1_NAjGkj7_0cRta1GdR4c5M&e=> 
> your input would be welcome if you have a few minutes to take a look 
> at this and comment on it.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub 
> <https://urldefense.proofpoint.com/v2/url?u=https-3A__github.com_javafxports_openjdk-2Djfx_pull_424-23issuecomment-2D479178579&d=DwMCaQ&c=RoP1YumCXCgaWHvlZYR8PZh8Bv7qIrMUB65eapI_JnE&r=H6jKHjHoLf6vCNF0PYEnJ6cS6kScbGLraJqgoRs9znU&m=TMsfvQuwAKRQX0c5NNEKf5Oo1Fpp1ZYnCUAIfdHcnTs&s=BRfeqr7uAcpilMrmcbiraC6uj1vgGJMFDM69AFKb5S4&e=>, 
> or mute the thread 
> <https://urldefense.proofpoint.com/v2/url?u=https-3A__github.com_notifications_unsubscribe-2Dauth_AENU8qaiKXyTdtOMN4zczm2c2aKp5Y1kks5vc7k3gaJpZM4cSjw-2D&d=DwMCaQ&c=RoP1YumCXCgaWHvlZYR8PZh8Bv7qIrMUB65eapI_JnE&r=H6jKHjHoLf6vCNF0PYEnJ6cS6kScbGLraJqgoRs9znU&m=TMsfvQuwAKRQX0c5NNEKf5Oo1Fpp1ZYnCUAIfdHcnTs&s=E2HNqem6K6aZCq6tDZ6GxaW5u0fy83WFufIxiVG8lkg&e=>.
>