@dgrieve 2019-04-03
Did you consider changing the lookupFont method such that it never returns SKIP?  lookupFont is an expensive little bugger since it considers all font, font-weight, font-size, and font-style styles up to the root (if necessary). If it returns SKIP, then (worse case) the code goes about calling getCachedFont on parents. If lookupFont returns SKIP and the parent's font isn't cached, then won't parent.lookupFont  also return SKIP? So perhaps lookupFont should always cache a font for the given pseudoclass state and never return SKIP.